Detect correct path for output
If the input was in a different directory than the genconfig
executable, the output would be placed in the executable's
directory and not the config directory.

Note that if the output is explicitly specified, it must be
in the same package as the input config in order to be able to
read the config. A good location won't be detected automatically.

Closes #3