Fix dotenv generation and build tags
The main function accidentally passed the default generated dotenv
location as the generated dotenv argument, instead of the one supplied
from the command line, if any was provided. It also used the command
line dotenv value as the value passed for the argument that adds build
tags, which is currently used for testcases. This essentially meant that
genconfig was using the dotenv output value as build tags, breaking the
generation.